refactor: move engine core files
